New module for switzerland specific sponsorship modifications.
