+++ sponsorship_switzerland/view/contracts_view.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+    Copyright (C) 2015 Compassion (http://www.compassion.ch)
+    @author Emanuel Cino <ecino@compassion.ch>
+    The licence is in the file __openerp__.py
+-->
+<openerp>
+	<data>
+        <!-- Search Filters -->
+        <record id="contract_tracking_switzerland_filter" model="ir.ui.view">
+			<field name="name">contract.tracking.switzerland.select</field>
+			<field name="model">recurring.contract</field>
+            <field name="inherit_id" ref="sponsorship_tracking.contract_tracking_filter"/>
+			<field name="arch" type="xml">
+                <field name="project_state" position="after">
+                    <field name="gmc_state"/>
+                </field>
+                <xpath expr="//filter[@string='Project Status']" position="after">
+                    <filter string="GMC Update" domain="[]"  context="{'group_by':'gmc_state'}"/>
+                </xpath>
+            </field>
+        </record>
+        
+        <record id="view_recurring_contract_compassion_filter" model="ir.ui.view">
+			<field name="name">recurring.contract.compassion.select</field>
+			<field name="model">recurring.contract</field>
+			<field name="inherit_id" ref="message_center_compassion.view_gmc_contract_filter" />
+			<field name="arch" type="xml">
+				<data>
+                    <filter name="biennial" position="before">
+                        <filter string="Order picture" domain="[('gmc_state', '=', 'order_picture')]"/>
+                    </filter>
+                </data>
+            </field>
+        </record>
+    </data>
+</openerp>
